requirements.txt
aiohttp==3.9.1
requests==2.31.0
pyyaml==6.0.1
python-telegram-bot==20.7
asyncio==3.4.3
colorama==0.4.6
